downloader-py/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
httpx==0.26.0
pycryptodome==3.19.0
python-multipart==0.0.6
pydantic==2.5.2
aiofiles==23.2.1